rips the ui files from TG for the library console
